Move LanguageSupportable on ResourceModel into separate file extension
